backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.3.4
nltk==3.8.1
# Using pre-built wheels to avoid compilation issues
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
# Simplified NLP dependencies
textblob==0.17.1
